requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pypdf
python-docx
openai

textract
pywin32
numpy
